isolate interface definition
remove all implementation code and add mock tests for the interface
portion of the assigment
