--- CMakeLists.txt
@@ -1,59 +1,0 @@
-# Set the minimum CMake version and policies for highest tested version
-cmake_minimum_required(VERSION 3.30)
-
-# Set up the project and ensure there is a working C++ compiler
-project(python_graph_library LANGUAGES CXX)
-find_package(Boost REQUIRED COMPONENTS graph system filesystem)
-include_directories(${Boost_INCLUDE_DIRS})  
-
-
-# Warn if the user invokes CMake directly
-if (NOT SKBUILD)
-  message(WARNING "\
-  This CMake file is meant to be executed using 'scikit-build-core'.
-  Running it directly will almost certainly not produce the desired
-  result. If you are a user trying to install this package, use the
-  command below, which will install all necessary build dependencies,
-  compile the package in an isolated environment, and then install it.
-  =====================================================================
-   $ pip install .
-  =====================================================================
-  If you are a software developer, and this is your own package, then
-  it is usually much more efficient to install the build dependencies
-  in your environment once and use the following command that avoids
-  a costly creation of a new virtual environment at every compilation:
-  =====================================================================
-   $ pip install nanobind scikit-build-core[pyproject]
-   $ pip install --no-build-isolation -ve .
-  =====================================================================
-  You may optionally add -Ceditable.rebuild=true to auto-rebuild when
-  the package is imported. Otherwise, you need to rerun the above
-  after editing C++ files.")
-endif()
-
-# Try to import all Python components potentially needed by nanobind
-find_package(Python 3.8
-  REQUIRED COMPONENTS Interpreter Development.Module
-  OPTIONAL_COMPONENTS Development.SABIModule)
-
-# Import nanobind through CMake's find_package mechanism
-find_package(nanobind CONFIG REQUIRED)
-
-# We are now ready to compile the actual extension module
-nanobind_add_module(
-  # Name of the extension
-  _booty_boost
-
-  # Target the stable ABI for Python 3.12+, which reduces
-  # the number of binary wheels that must be built. This
-  # does nothing on older Python versions
-  STABLE_ABI
-
-  # Source code goes here
-  src/booty_boost.cpp
-  src/Graph.hpp
-  src/Graph.cpp
-)
-
-# Install directive for scikit-build-core
-install(TARGETS _booty_boost LIBRARY DESTINATION booty_boost)
