--- test.py
@@ -1,3 +1,0 @@
-import booty_boost as bb
-G = bb.Graph()
-print(dir(G))  # This should print a list of methods including 'is_tree'
