+++ tests/sanity_test.py
@@ -1,0 +1,8 @@
+from unittest.mock import MagicMock
+import booty_boost as bb
+
+"""Simple script test to just make sure the project is building and class exists."""
+
+G = MagicMock(bb.BaseGraph)
+
+print(dir(G))  # This should print a list of methods including 'is_tree'
